feat! Added new HTML elements along with CSS and unfinished JavaScript
Added new button html class, with its repective css. Tweaked the
Added an unfinished JS if statement, will come back to it later
Positioned searchbar correctly, currently experimenting with
substituting <input type=submit> with <button>, since I am unsure of
how to merge the "search symbol"(&#9740) with the <input>, since it
defaults... to block.... If I added inline, would it work??
Will come back to that later.
Changed colours to fit that of a professional website's